Project support, reorganize files, stdout output
- Pass the project to executed rules in order to use the project path
as context
- Move functions around for clariry
- make the checker Run function return the projects results
