+++ .gitignore
@@ -1,1 +1,2 @@
 standards
+standards-insights
