+++ cmd/root.go
@@ -1,8 +1,18 @@
 package cmd
 
 import (
+	"fmt"
+	"os"
+
 	"github.com/spf13/cobra"
 )
+
+func exit(err error) {
+	if err != nil {
+		fmt.Println(err.Error())
+		os.Exit(2)
+	}
+}
 
 func Run() error {
 	configPath := ""
