+++ main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"fmt"
 	"os"
 
 	"github.com/qonto/standards-insights/cmd"
@@ -9,6 +10,7 @@ import (
 func main() {
 	err := cmd.Run()
 	if err != nil {
+		fmt.Println(err.Error())
 		os.Exit(1)
 	}
 }
